# feat: improve Analyzed performance on a large schema