json.Unmarshal is too heavy on a large schema.

```
name        old time/op  new time/op  delta
Analyzed-4   964ms ±11%   522ms ± 4%  -45.90%  (p=0.000 n=10+10)
```